Feature author expertise (#422)
Creation of `expertise` model and `expertisable` relationship - expertises can now be added to the author profiles. Expertise used are from the sciences.gc.ca and RESE lists. Keeping only unique values and discarding the taxonomy of the science profiles which introduced too much redundancy, contradictions, as well as a few typos.