refactor(author): update ManageAuthorOrcidCard.vue
This commit refactors the ManageAuthorOrcidCard.vue component by updating the code to use the new i18n module. It replaces the usage of the deprecated $t function with the new t function from the useI18n hook. This ensures that the component is using the latest i18n functionality and improves code maintainability.