feat: submit manuscript to single window
Authors must now attached their manuscript in word format when doing
their submisison to the single window team. This removes an additional
step fo having the single window reach out to the author to ask for it.
The file is also saved as a supplementary file in the publicaiton for
later reference.